+++ Xext/geext/geext.c
@@ -31,7 +31,7 @@
 #include "dix/dix_priv.h"
 #include "dix/request_priv.h"
 #include "miext/extinit_priv.h"
-#include "Xext/geext_priv.h"
+#include "geext_priv.h"
 
 #include "windowstr.h"
 #include "protocol-versions.h"
