+++ Xext/geext/meson.build
@@ -1,0 +1,9 @@
+srcs_geext = [
+    'geext.c',
+]
+
+libxserver_geext = static_library('xserver_geext',
+    srcs_geext,
+    include_directories: inc,
+    dependencies: common_dep,
+)
