+++ Xext/xinput/extinit.c
@@ -65,7 +65,7 @@ SOFTWARE.
 #include "dix/exevents_priv.h"
 #include "dix/extension_priv.h"
 #include "miext/extinit_priv.h"
-#include "Xext/geext_priv.h"
+#include "Xext/geext/geext_priv.h"
 
 #include "inputstr.h"
 #include "gcstruct.h"           /* pointer for extnsionst.h */
